+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Input box is now larger and wraps text to show more data
   - Adjusted input/delete buttons to stack on the right of the input widget
 - Disabled scroll wheel for language combo box
+- Improved the Output tab layout/visuals
 
 ### Fixed